move the client and Rx into the provider
